+++ test/specs/unit/utilSpec.js
@@ -204,23 +204,23 @@ describe("The Util Library", function() {
     });
 
 
-    describe("The setImmediate() method", function() {
+    describe("The defer() method", function() {
         it("Should call methods in the right order", function(done) {
             var result = [];
-            layer.Util.setImmediate(function() {
+            layer.Util.defer(function() {
                 result.push("a");
             });
-            layer.Util.setImmediate(function() {
+            layer.Util.defer(function() {
                 result.push("b");
             });
-            layer.Util.setImmediate(function() {
+            layer.Util.defer(function() {
                 result.push("c");
             });
-            layer.Util.setImmediate(function() {
+            layer.Util.defer(function() {
                 result.push("d");
             });
             result.push("before");
-            layer.Util.setImmediate(function() {
+            layer.Util.defer(function() {
                 expect(result).toEqual(["before", "a", "b", "c", "d"]);
                 setTimeout(function() {
                     done();
@@ -228,19 +228,19 @@ describe("The Util Library", function() {
             });
         });
 
-        it("Should run functions queued within setImmediate after all other queued calls", function(done) {
+        it("Should run functions queued within defer after all other queued calls", function(done) {
             var result = [];
 
-            layer.Util.setImmediate(function() {
+            layer.Util.defer(function() {
                 result.push(1);
-                layer.Util.setImmediate(function() {
+                layer.Util.defer(function() {
                     result.push(11);
                 });
             });
 
-            layer.Util.setImmediate(function() {
+            layer.Util.defer(function() {
                 result.push(2);
-                layer.Util.setImmediate(function() {
+                layer.Util.defer(function() {
                     result.push(22);
                 });
             });