+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.0.3
 
 * Adds a `setImmediate` function that defers execution without using `setTimeout` but uses a `setTimeout` fallback.
+* Fixes the `move` event for `layer.Query`.
 
 ## 3.0.2